+++ templates/personal.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+use OCP\Util;
+
+Util::addScript('roomvox', 'roomvox-personal');
+Util::addStyle('roomvox', 'style');
+?>
+
+<div id="app-roomvox-personal"></div>